ALSA: emu10k1: introduce alternative E-MU D.A.S. mode
As noted in a previous commit, the E-MU Digital Audio System cards
don't try very hard to be Sound Blasters. This commit takes it further
and introduces a module option to switch to a completely separate mode.

In that mode:
- The regular PCM playback/capture devices are removed
- The multi-channel playback/capture devices get index 0
- Consequently, the regular mixer controls are also completely removed,
  as the multi-channel devices already bypassed them.
  This is no real loss, given the expected use with a sound server.
- The voice send routing+amount & attenuation controls are also removed,
  as they are just another mixer. The routing is redundant with the FPGA
  channel routing, and amounts/att'n can be done in software. The
  hardware implementation of the latter is also incompatible with 32-bit
  playback, which we'll add support for later.
- A_EXTOUT is now also free for multi-channel capture, so we use that
  instead of A_FXBUS2 - this will later simplify using both at once.
- For the same reason, the FX outputs are not listed in /proc anymore
- The device name is changed, so mixer settings don't get mixed up

This continues the pre-existing design with a single multi-channel
device where the channels are routed by manual mixer controls from/to
the physical ports. De-/multiplexing must be done by the sound server
if independent streams are to be used.

An alternative design would be exposing each physical port as a separate
device and automatically setting up the routes. This would be a somewhat
radical departure from the status quo, and I don't know whether it would
be a net benefit. It certainly would be harder to implement, as it would
require sync start of streams and a channel allocator (the latter would
have the added benefit of properly reporting EMU32 & EDI bus over-
allocation, which the mixer does not). And there is only one big multi-
channel capture engine, so streams would have to be de-multiplexed by
the driver in software - which seems a bit counter-productive if the
sound server would re-multiplex them again.

Signed-off-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Change-Id: Idc4b24135eca33022cf59abb21b3112353dff85a