ALSA: emu10k1: do away with extra voices in E-MU D.A.S. mode
Maybe somewhat unsurprisingly, Creative Inc. learned from the period
interrupt disaster of the EMU10K1, and equipped Audigy with a way to
delay the interrupts by a configurable amount. So make use of that.
This allows us to do away with the extra voices for driving interrupts.

The most salient advantage of that is that we can maximize the number of
playback channels in 24-bit 192 kHz mode, bringing it up to 8.
A minor point is that no bandwidth is wasted on fetching data that is
discarded anyway.

The removal of the extra voices has one disadvantage: the hardware is
limited to two periods, thus increasing the likelihood of underruns.
Therefore, we do this only in D.A.S. mode, which aims at high-quality
low-latency multi-channel recording, and therefore can be expected to be
operated in an optimal environment. Also, prior to a few commits back,
the multi-channel playback was actually limited to two periods anyway
(for no good reason).

Signed-off-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Change-Id: I198750a02807ce62c0abe5a070aeed8a557ddba6