ASoC: au1x: declare that FIFO is reported in frames
Signed-off-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
---

Is this right? The comment on the field says "fifo entries of AC97/I2S
PSC", which doesn't suggest bytes. The data sheet speaks of "words" and
"byte masks", but without digging into it I can't tell how it would
behave with different sample widths and channel counts (which the driver
does not seem to limit _at all_? what am I missing?).

This is just something I stumbled over while surveying how different
drivers set the field. If this is actually a bug, it is minor.

Change-Id: I240c4acbef28bdb26fcee70934f5e29841ef0341