ALSA: emu10k1: add support for 2x/4x word clocks in E-MU D.A.S. mode
This lays the groundwork for supporting 88.2/96/176.4/192 kHz rates
without actually doing so yet - we simply multi-feed the same samples
on playback, and throw away the excess ones on capture. Input-to-output
monitoring does actually use the full sample rate, though.

Due to hardware constraints, changing the clock multiplier (CM) changes
the available audio ports and the number of available channels. This has
an impact on the channel routing mixer controls. One way to deal with
this would be presenting a union of all possibilities, and simply
ignoring currently inapplicable settings. However, this would be a
terrible user experience, and go against the spirit of prior patches
aimed at decluttering the mixer. Therefore, we do dynamic
reconfiguration (DR) of the mixer in response to changing the CM.

DR is somewhat controversial, as it has the potential to crash poorly
programmed applications. But that in itself isn't a very convincing
argument against it, as by that logic we'd have to ban all hot-plugging.
Such crashes would also not really qualify as regressions, as the D.A.S.
mode is a new opt-in feature, and therefore no previously stable setups
would be impacted. Also, pendantically, the driver already had DR via
SNDRV_EMU10K1_IOCTL_CODE_POKE. A somewhat valid concern is that changing
mixer settings is a non-privileged operation and therefore potential
crashes could be exploited for a somewhat more impactful nuisance attack
on another user than messing with the mixer per se. However, systemd &
co. limit device access to the user currently logged in on the seat
owning the device.

There is a specific concern about doing DR in response to changing a
mixer control's value, as an application may legitimately react to DR by
updating all mixer settings in turn. However, that update should write
the same value to the clock multiplier, thus terminating the recursion.

One may limit DR to merely disabling inapplicable controls, in the hope
that this would be better handled than completely tearing down and
re-creating controls as we do. However, there is no guarantee for that.
And because it is impossible to disable particular enum values within a
control, it would be necessary to have three complete sets of
per-channel controls. This would yield an extremely cluttered and
confusing UI if the application (reasonably) chose to merely visually
disable inactive controls rather than hiding them.

We do the DR synchronously from within snd_emu1010_clock_shift_put().
This was enabled by commit 5bbb1ab5bd ("control: use counting semaphore
as write lock for ELEM_WRITE operation"); we merely need to make
add_ctls() use snd_ctl_add_locked() instead of snd_ctl_add(), so it
doesn't deadlock.

It would be possible to do the DR in a tasklet after the ioctl finishes.
However, it is not obvious what actual problem that would solve, and the
added asynchronicity would significantly complicate matters, esp. wrt.
the batch updates expected during mixer state restoration.

Signed-off-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Change-Id: I89c38dc5cc0f14c2b2886803ff493b693c230c57