ALSA: emu10k1: limit external clock sources on E-MU 1616 CardBus
Turns out that the card offers the MicroDock's external clock inputs
both as a multiplexed "Dock" input and as separate S/PDIF and ADAT
inputs, so neither option is dysfunctional. But for simplicity, we offer
only the "Dock" option.

Also, remove the BNC option, as the card has no sync card option - that
was a mistake in 60985241bfc6i (ALSA: emu10k1: make available E-MU clock
sources card-specific).

Signed-off-by: Oswald Buddenhagen <oswald.buddenhagen@gmx.de>
Change-Id: Id3099d16a48c84a90b07f31684294cc6925c0d47